bumscan_backend/pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["D4reD3VIL81 <ahmad.81.sadeghi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^24.1"
django-cors-headers = "^4.3.1"
whitenoise = "^6.6.0"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
